Convert content of MainActivity to fragment
