+++ app/src/main/res/layout/activity_main.xml
@@ -27,7 +27,11 @@
 
 		</com.google.android.material.appbar.AppBarLayout>
 
-		<include layout="@layout/activity_main_content"/>
+		<FrameLayout
+			android:id="@+id/content_main"
+			android:layout_width="match_parent"
+			android:layout_height="match_parent"
+			app:layout_behavior="@string/appbar_scrolling_view_behavior" />
 
 	</androidx.coordinatorlayout.widget.CoordinatorLayout>
