+++ app/src/main/res/layout/fragment_infocenter.xml
@@ -1,11 +1,9 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
 	xmlns:android="http://schemas.android.com/apk/res/android"
-	xmlns:app="http://schemas.android.com/apk/res-auto"
 	android:id="@+id/swiperefreshlayout_infocenter"
 	android:layout_width="match_parent"
-	android:layout_height="match_parent"
-	app:layout_constraintTop_toTopOf="parent">
+	android:layout_height="match_parent">
 
 	<androidx.recyclerview.widget.RecyclerView
 		android:id="@+id/recyclerview_infocenter"
