+++ app/src/main/res/layout/fragment_timetable.xml
@@ -2,12 +2,8 @@
 <androidx.constraintlayout.widget.ConstraintLayout
 	xmlns:android="http://schemas.android.com/apk/res/android"
 	xmlns:app="http://schemas.android.com/apk/res-auto"
-	xmlns:tools="http://schemas.android.com/tools"
-	android:id="@+id/content_main"
 	android:layout_width="match_parent"
-	android:layout_height="match_parent"
-	app:layout_behavior="@string/appbar_scrolling_view_behavior"
-	tools:context=".activities.MainActivity">
+	android:layout_height="match_parent">
 
 	<com.sapuseven.untis.views.WeekViewSwipeRefreshLayout
 		android:id="@+id/swiperefreshlayout_main_timetable"
